Content Profile Support for Extensions
Adds support for "mustSupports" in extensions based on the ContentProfile
