+++ lib/common.js
@@ -296,6 +296,122 @@ class TargetItem {
   }
 }
 
+class FieldTarget {
+  constructor(target, commands=[], comments) {
+    this._target = target;
+    this._commands = commands;
+    this._comments = comments;
+  }
+
+  static parse(ruleTarget) {
+    const matches = /([^\s\(]+)(\s+\((.+)\))?(\s+\/\/\s*(.*))?/.exec(ruleTarget);
+    if (matches == null || typeof matches[1] === 'undefined') {
+      return;
+    }
+    const target = matches[1];
+    let commands, comments;
+    if (typeof matches[3] !== 'undefined') {
+      commands = MappingCommand.parseMany(matches[3]);
+    }
+    if (typeof matches[5] !== 'undefined') {
+      comments = matches[5];
+    }
+    return new FieldTarget(target, commands, comments);
+  }
+
+  get target() { return this._target; }
+  get commands() { return this._commands; }
+  get comments() { return this._comments; }
+
+  // Functions to check if the target is an extension and how its mapped (URL or element)
+  isExtensionURL() {
+    return /^https?:\/\//.test(this._target);
+  }
+  isExtensionPath() {
+    return /^(.+\.)?(extension|modifierExtension)$/.test(this._target);
+  }
+  isExtension() {
+    return this.isExtensionURL() || this.isExtensionPath();
+  }
+
+  // The "slice at" command indicates the path where the slicing is rooted.  This is only needed when the target
+  // path is *not* where the root of the slice should be (for example, if the target is not multiple cardinality).
+  hasSliceAtCommand() {
+    return this._commands.some(c => c.key == 'slice at');
+  }
+  findSliceAtCommand() {
+    return this._commands.find(c => c.key == 'slice at');
+  }
+  addSliceAtCommand(at) {
+    this._commands.push(new MappingCommand('slice at', at));
+  }
+
+  // The "slice on" command indicates what FHIR calls the "discriminator path".
+  hasSliceOnCommand() {
+    return this._commands.some(c => c.key == 'slice on');
+  }
+  findSliceOnCommand() {
+    return this._commands.find(c => c.key == 'slice on');
+  }
+  addSliceOnCommand(on) {
+    this._commands.push(new MappingCommand('slice on', on));
+  }
+
+  // The "slice on type" command indicates what FHIR calls the "discriminator type".
+  // If not set, the "value" type is typically used.
+  hasSliceOnTypeCommand() {
+    return this._commands.some(c => c.key == 'slice on type');
+  }
+  findSliceOnTypeCommand() {
+    return this._commands.find(c => c.key == 'slice on type');
+  }
+  addSliceOnTypeCommand(on) {
+    this._commands.push(new MappingCommand('slice on type', on));
+  }
+
+  // The "in slice" command is for elements to indicate what slice they belong to (by slice name).
+  // This is not typically set in the mapping file, but rather, applied by the shr-fhir-export logic.
+  hasInSliceCommand() {
+    return this._commands.some(c => c.key == 'in slice');
+  }
+  findInSliceCommand() {
+    return this._commands.find(c => c.key == 'in slice');
+  }
+  addInSliceCommand(sliceName) {
+    this._commands.push(new MappingCommand('in slice', sliceName));
+  }
+
+  // The "slice #" command is used to map elements to an existing slice (when mapping to a profile).
+  // We use slice # since slice name isn't easily exposed in IG documentation.  Slice # starts with 1.
+  hasSliceNumberCommand() {
+    return this._commands.some(c => c.key == 'slice #');
+  }
+  findSliceNumberCommand() {
+    return this._commands.find(c => c.key == 'slice #');
+  }
+  addSliceNumberCommand(sliceNumber) {
+    this._commands.push(new MappingCommand('slice #', sliceNumber));
+  }
+
+  // The "slice strategy" command currently only supports one strategy: "includes".  If not set, then it slices
+  // based on mappings that share target paths.
+  hasSliceStrategyCommand() {
+    return this._commands.some(c => c.key == 'slice strategy');
+  }
+  findSliceStrategyCommand() {
+    return this._commands.find(c => c.key == 'slice strategy');
+  }
+  addSliceStrategyCommand(strategy) {
+    this._commands.push(new MappingCommand('slice strategy', strategy));
+  }
+
+  toRuleTarget() {
+    const commandStr = this._commands.length == 0 ? '' : ` (${this._commands.map(c => c.toString()).join('; ')})`;
+    const commentsStr = typeof this._comments === 'undefined' ? '' : ` // ${this._comments}`;
+    return `${this._target}${commandStr}${commentsStr}`;
+  }
+}
+
 class MappingCommand {
   constructor(key, value) {
     this._key = key;
@@ -327,4 +443,4 @@ class FHIRExportError extends Error {
   }
 }
 
-module.exports = {FHIRExportError, getSnapshotElement, getSnapshotElementById, getDifferentialElementById, hasSlicingOnBaseElement, addSlicingToBaseElement, createSlicingObject, elementTypeContainsTypeName, fhirID, fhirURL, shortID, valueAndFields, valueName, equalShrElementPaths, escapeHTML, cloneJSON, capitalize, lowerFirst, todayString, isCustomProfile, typeToString, trim, getTarget, ProcessTracker, TargetItem, MappingCommand};
+module.exports = {FHIRExportError, getSnapshotElement, getSnapshotElementById, getDifferentialElementById, hasSlicingOnBaseElement, addSlicingToBaseElement, createSlicingObject, elementTypeContainsTypeName, fhirID, fhirURL, shortID, valueAndFields, valueName, equalShrElementPaths, escapeHTML, cloneJSON, capitalize, lowerFirst, todayString, isCustomProfile, typeToString, trim, getTarget, ProcessTracker, TargetItem, FieldTarget, MappingCommand};
