+++ lib/export.js
@@ -3,6 +3,7 @@ const mdls = require('shr-models');
 const load = require('./load');
 const common = require('./common');
 const MVH = require('./multiVersionHelper');
+const {FieldTarget} = require('./common');
 const {insertElementInSnapshot, insertElementInDifferential} = require ('./insertElement');
 const {CodeSystemExporter} = require('./codeSystems');
 const {ValueSetExporter} = require('./valueSets');
@@ -3170,54 +3171,6 @@ class FHIRExporter {
     }
   }
 
-  /**
-   * Inserts an element into its proper place in a list of elements
-   */
-  insertElementInList(element, list) {
-    let i = 0;
-    let lastMatchId = '';
-    for (; i < list.length; i++) {
-      const currentId = list[i].id;
-      // If the item we're placing starts with the current item, it's a match
-      // so remember the match and go to the next element in the list
-      if ((new RegExp(`^${escapeRegExp(currentId)}([\\.:].+)?$`)).test(element.id)) {
-        lastMatchId = currentId;
-      // If it wasn't a match, but the current item is a choice (e.g., value[x]), call it
-      // a match if it starts with the same root (e.g., valueString), and then go to the next element
-      } else if (currentId.endsWith('[x]') && (new RegExp(`^${escapeRegExp(currentId.slice(0, -3))}[A-Z][^\\.]+(\\..+)?$`)).test(element.id)) {
-        lastMatchId = currentId;
-      } else {
-        let stop;
-        // If the next part of the item is a '.' (not a ':'), then we want to stop if the current item
-        // doesn't start with the last match or is a slice of the last match (as indicated by a ':')
-        if (element.id.length > lastMatchId.length && element.id[lastMatchId.length] === '.') {
-          stop = ! new RegExp(`^${escapeRegExp(lastMatchId)}(\\..+)?$`).test(currentId);
-        // else we want to stop if the current item doesn't start with the last match
-        } else {
-          stop = ! new RegExp(`^${escapeRegExp(lastMatchId)}([\\.:].+)?$`).test(currentId);
-        }
-        if (stop) {
-          break;
-        }
-      }
-    }
-    list.splice(i, 0, element);
-  }
-
-  /**
-   * Inserts an element into its proper place in the snapshot element
-   */
-  insertElementInSnapshot(element, profile) {
-    this.insertElementInList(element, profile.snapshot.element);
-  }
-
-  /**
-   * Inserts an element into its proper place in the differential elements
-   */
-  insertElementInDifferential(element, profile) {
-    this.insertElementInList(element, profile.differential.element);
-  }
-
   lookupProfile(identifier, createIfNeeded=true, warnIfProfileIsProcessing=false) {
     let p = this._profilesMap.get(common.fhirID(identifier));
     if (typeof p === 'undefined' && createIfNeeded) {
@@ -3888,123 +3841,6 @@ class FHIRExporter {
     const [res, path] = el.path.split('.', 2);
     const resJSON = this._fhir.find(res);
     return common.getSnapshotElement(resJSON, path);
-  }
-}
-
-// NOTE: Similar "TargetItem" class is in common.js
-class FieldTarget {
-  constructor(target, commands=[], comments) {
-    this._target = target;
-    this._commands = commands;
-    this._comments = comments;
-  }
-
-  static parse(ruleTarget) {
-    const matches = /([^\s\(]+)(\s+\((.+)\))?(\s+\/\/\s*(.*))?/.exec(ruleTarget);
-    if (matches == null || typeof matches[1] === 'undefined') {
-      return;
-    }
-    const target = matches[1];
-    let commands, comments;
-    if (typeof matches[3] !== 'undefined') {
-      commands = common.MappingCommand.parseMany(matches[3]);
-    }
-    if (typeof matches[5] !== 'undefined') {
-      comments = matches[5];
-    }
-    return new FieldTarget(target, commands, comments);
-  }
-
-  get target() { return this._target; }
-  get commands() { return this._commands; }
-  get comments() { return this._comments; }
-
-  // Functions to check if the target is an extension and how its mapped (URL or element)
-  isExtensionURL() {
-    return /^https?:\/\//.test(this._target);
-  }
-  isExtensionPath() {
-    return /^(.+\.)?(extension|modifierExtension)$/.test(this._target);
-  }
-  isExtension() {
-    return this.isExtensionURL() || this.isExtensionPath();
-  }
-
-  // The "slice at" command indicates the path where the slicing is rooted.  This is only needed when the target
-  // path is *not* where the root of the slice should be (for example, if the target is not multiple cardinality).
-  hasSliceAtCommand() {
-    return this._commands.some(c => c.key == 'slice at');
-  }
-  findSliceAtCommand() {
-    return this._commands.find(c => c.key == 'slice at');
-  }
-  addSliceAtCommand(at) {
-    this._commands.push(new common.MappingCommand('slice at', at));
-  }
-
-  // The "slice on" command indicates what FHIR calls the "discriminator path".
-  hasSliceOnCommand() {
-    return this._commands.some(c => c.key == 'slice on');
-  }
-  findSliceOnCommand() {
-    return this._commands.find(c => c.key == 'slice on');
-  }
-  addSliceOnCommand(on) {
-    this._commands.push(new common.MappingCommand('slice on', on));
-  }
-
-  // The "slice on type" command indicates what FHIR calls the "discriminator type".
-  // If not set, the "value" type is typically used.
-  hasSliceOnTypeCommand() {
-    return this._commands.some(c => c.key == 'slice on type');
-  }
-  findSliceOnTypeCommand() {
-    return this._commands.find(c => c.key == 'slice on type');
-  }
-  addSliceOnTypeCommand(on) {
-    this._commands.push(new common.MappingCommand('slice on type', on));
-  }
-
-  // The "in slice" command is for elements to indicate what slice they belong to (by slice name).
-  // This is not typically set in the mapping file, but rather, applied by the shr-fhir-export logic.
-  hasInSliceCommand() {
-    return this._commands.some(c => c.key == 'in slice');
-  }
-  findInSliceCommand() {
-    return this._commands.find(c => c.key == 'in slice');
-  }
-  addInSliceCommand(sliceName) {
-    this._commands.push(new common.MappingCommand('in slice', sliceName));
-  }
-
-  // The "slice #" command is used to map elements to an existing slice (when mapping to a profile).
-  // We use slice # since slice name isn't easily exposed in IG documentation.  Slice # starts with 1.
-  hasSliceNumberCommand() {
-    return this._commands.some(c => c.key == 'slice #');
-  }
-  findSliceNumberCommand() {
-    return this._commands.find(c => c.key == 'slice #');
-  }
-  addSliceNumberCommand(sliceNumber) {
-    this._commands.push(new common.MappingCommand('slice #', sliceNumber));
-  }
-
-  // The "slice strategy" command currently only supports one strategy: "includes".  If not set, then it slices
-  // based on mappings that share target paths.
-  hasSliceStrategyCommand() {
-    return this._commands.some(c => c.key == 'slice strategy');
-  }
-  findSliceStrategyCommand() {
-    return this._commands.find(c => c.key == 'slice strategy');
-  }
-  addSliceStrategyCommand(strategy) {
-    this._commands.push(new common.MappingCommand('slice strategy', strategy));
-  }
-
-  toRuleTarget() {
-    const commandStr = this._commands.length == 0 ? '' : ` (${this._commands.map(c => c.toString()).join('; ')})`;
-    const commentsStr = typeof this._comments === 'undefined' ? '' : ` // ${this._comments}`;
-    return `${this._target}${commandStr}${commentsStr}`;
   }
 }
