+++ lib/extensions.js
@@ -94,6 +94,11 @@ class ExtensionExporter {
     this.pushExtensionBaseElement(extension, baseId, def, card);
     this.pushExtensionIdElement(extension, baseId);
 
+    const cp = this._specs.contentProfiles.findByIdentifier(def.identifier);
+    if (cp != null) {
+      console.log('EXT CONTENT PROFILE:', cp.identifier.fqn);
+    }
+
     // If this supports a simple extension, then determine its type (and other necessary data)
     let type, value;
     const map = this._specs.maps.findByTargetAndIdentifier(this._target, def.identifier);
@@ -176,7 +181,7 @@ class ExtensionExporter {
     } else {
       this.pushExtensionNoExtensionsElement(extension, baseId);
       this.pushExtensionURLElement(extension, baseId, def);
-      this.pushExtensionValueXElement(extension, baseId, type, value);
+      this.pushExtensionValueXElement(extension, baseId, def, type, value);
     }
   }
 
@@ -314,7 +319,7 @@ class ExtensionExporter {
     extension.differential.element.push(dfURL);
   }
 
-  pushExtensionValueXElement(extension, baseId, type, value) {
+  pushExtensionValueXElement(extension, baseId, def, type, value) {
     const ucType = type.charAt(0).toUpperCase() + type.slice(1);
     const ssValue = {
       id: `${baseId}.value${ucType}`,
@@ -366,6 +371,55 @@ class ExtensionExporter {
     }
     if (value instanceof mdls.IdentifiableValue) {
       this._profileExporter.applyConstraints(value, extension, ssValue, dfValue, false);
+    }
+
+    const cp = this._specs.contentProfiles.findByIdentifier(def.identifier);
+    if (cp != null) {
+      const cpRules = cp.rules.filter(r => {
+        if (r.path[0].isValueKeyWord) {
+          return true;
+        } else if (value instanceof mdls.IdentifiableValue) {
+          return r.path[0].equals(value.effectiveIdentifier);
+        } else if (value instanceof mdls.ChoiceValue) {
+          return value.aggregateOptions.some(opt => r.path[0].equals(opt));
+        }
+      });
+      cpRules.forEach(r => {
+        if (r.path.length === 1) {
+          console.log(`  MATCHED ${r.path.map(id => id.name).join('.')}: ${ssValue.id}`);
+          if (r.mustSupport) {
+            ssValue.mustSupport = dfValue.mustSupport = true;
+          }
+        } else {
+          // Partial match: dive further into the element to get the full match.
+          // Get the expected FHIR subpath for the remaining identifier path
+          const targetSubPath = this._profileExporter.findTargetFHIRPath(r.path[0], r.path.slice(1));
+          if (targetSubPath) {
+            // The subpath was determined, so now get the "Value" for the tail of the path,
+            // which is needed when getting the corresponding snapshot element
+            const def = this._specs.dataElements.findByIdentifier(r.path[0]);
+            const tailValue = this._profileExporter.findValueByPath(r.path.slice(1), def, false, []);
+            // Get the deeper snapshot elements for each of the matched elements
+            const targetPath = `${ssValue.path}.${targetSubPath}`;
+            // Convert to a field target so we can use an existing function
+            const fieldTarget = new common.FieldTarget(targetPath.slice(targetPath.indexOf('.') + 1)); // Remove resource prefix
+            const targetSS = this._profileExporter.getSnapshotElementForFieldTarget(extension, fieldTarget, tailValue);
+            if (targetSS) {
+              console.log(`  MATCHED ${r.path.map(id => id.name).join('.')}: ${targetSS.id}`);
+              if (r.mustSupport) {
+                const targetDF = common.getDifferentialElementById(extension, targetSS.id, true);
+                targetSS.mustSupport = targetDF.mustSupport = true;
+              }
+            } else {
+              logger.error('Could not find FHIR element with path %s for content profile rule with path %s.  ERROR_CODE:???',
+                targetPath, r.path.map(p => p.name).join('.'));
+            }
+          } else {
+            logger.error('Could not find FHIR element for content profile rule with path %s.  ERROR_CODE:???',
+                r.path.map(p => p.name).join('.'));
+          }
+        }
+      });
     }
   }
 
@@ -450,6 +504,41 @@ class ExtensionExporter {
     extension.snapshot.element.push(ssExt);
     extension.differential.element.push(dfExt);
     this._profileExporter.applyConstraints(field, extension, ssExt, dfExt, true);
+
+    const cp = this._specs.contentProfiles.findByIdentifier(def.identifier);
+    if (cp != null) {
+      const cpRules = cp.rules.filter(r => {
+        if (r.path[0].isValueKeyWord) {
+          // Check if the field is the value or an option of the value choice (if applicable)
+          const options = def.value instanceof mdls.ChoiceValue ? def.value.aggregateOptions : [def.value];
+          return options.some(opt => opt.effectiveIdentifier && opt.effectiveIdentifier.equals(field.effectiveIdentifier));
+        } else if (field instanceof mdls.IdentifiableValue) {
+          return r.path[0].equals(field.effectiveIdentifier);
+        } // that's it; current code paths won't ever allow a choice value to get into this logic
+      });
+      cpRules.forEach(r => {
+        if (r.path.length === 1) {
+          console.log(`  MATCHED ${r.path.map(id => id.name).join('.')}: ${ssExt.id}`);
+          if (r.mustSupport) {
+            ssExt.mustSupport = dfExt.mustSupport = true;
+          }
+        } else {
+          // Partial match: dive further into the element to get the full match.
+          // Get the target element for the remaining identifier path
+          const targetSS = this._profileExporter.getElementInExtension(r.path, extension, ssExt);
+          if (targetSS) {
+            console.log(`  MATCHED ${r.path.map(id => id.name).join('.')}: ${targetSS.id}`);
+            if (r.mustSupport) {
+              const targetDF = common.getDifferentialElementById(extension, targetSS.id, true);
+              targetSS.mustSupport = targetDF.mustSupport = true;
+            }
+          } else {
+            logger.error('Could not find FHIR element subextension for content profile rule with path %s.  ERROR_CODE:???',
+              r.path.map(p => p.name).join('.'));
+          }
+        }
+      });
+    }
   }
 
   getExtensionPathFromExtensionID(id) {
